requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.10.0
anthropic==0.34.0
python-dotenv==1.0.0
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.27.2
